adds update boarding for compatibility with germany
