+++ DP3TApp/Screens/Onboarding/NSOnboardingContentViewController.swift
@@ -16,7 +16,7 @@ class NSOnboardingContentViewController: NSViewController {
     private let defaultAnimationOffset: CGFloat = 200
     private let imageAnimationOffset: CGFloat = 400
 
-    var continueButtonText = "onboarding_continue_button".ub_localized
+    internal var continueButtonText = "onboarding_continue_button".ub_localized
 
     override func viewDidLoad() {
         super.viewDidLoad()
