+++ Translations/rm.lproj/InfoPlist.strings
@@ -1,3 +1,3 @@
 "CFBundleDisplayName" = "SwissCovid";
-"NSBluetoothPeripheralUsageDescription" = "Inscunters vegnan identifitgadas via bluetooth, economisant l'energia.";
-"NSBluetoothAlwaysUsageDescription" = "Inscunters vegnan identifitgadas via bluetooth, economisant l'energia.";
+"NSBluetoothPeripheralUsageDescription" = "Inscunters vegnan identifitgads via bluetooth, economisant l'energia.";
+"NSBluetoothAlwaysUsageDescription" = "Inscunters vegnan identifitgads via bluetooth, economisant l'energia.";
